feat: move unixtime into own module so can be imported in timestamp
