+++ go.mod
@@ -30,6 +30,7 @@ replace (
 	crabot/formatting => ./lib/formatting
 	crabot/formatting/markdown => ./lib/formatting/markdown
 	crabot/timestamp => ./lib/timestamp
+	crabot/timestamp/unixtime => ./lib/timestamp/unixtime
 	internal/dicecmd => ./internal/dicecmd
 	internal/env => ./internal/env
 )
