+++ go.work
@@ -8,6 +8,7 @@ use (
 	./lib/formatting
 	./lib/formatting/markdown
 	./lib/timestamp
+	./lib/timestamp/unixtime
 
 	./internal/env
 )
