+++ lib/timestamp/unixtime/unixTime.go
@@ -1,4 +1,4 @@
-package timestamp
+package unixtime
 
 import (
 	"errors"
